FUSETOOLS-2684 - extract data structure for TemplateItem
it avoids to have constructor with too many parameters.
